build 되도록 Long -> Integer 로 변경 및 dependency 추가

+++ src/main/java/zerobase/matching/chat/dto/ChatDto.java
@@ -10,8 +10,8 @@ import zerobase.matching.chat.type.ChatType;
 @Builder
 public class ChatDto {
 
-    private Long userId;
-    private Long roomId;
+    private int userId;
+    private int roomId;
     private String sender;
     private String chatContext;
     private ChatType chatType;
